Module Adaptive_issuance_storage.For_RPC

val get_reward_coeff : Raw_context.t -> cycle:Cycle_repr.t -> Tezos_protocol_environment_020_PsParisC.Q.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

get_reward_coeff ctxt cycle reads the reward coeff for the given cycle from the storage.

Fails if the given cycle is not between current_cycle and current_cycle + consensus_rights_delay.

If adaptive issuance has not been activated, then this function returns Q.one. Used only for RPCs. To get the actual rewards, use Delegate_rewards.

val get_reward_bonus : Raw_context.t -> cycle:Cycle_repr.t option -> Issuance_bonus_repr.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

get_reward_bonus ctxt cycle reads the reward bonus for the given cycle from the storage. If cycle is None, returns 0.

Returns 0 if the given cycle is not between current_cycle and current_cycle + consensus_rights_delay.

If adaptive issuance has not been activated, then this function returns 0. Used only for RPCs. To get the actual rewards, use Delegate_rewards.