Module Tezos_raw_protocol_020_PsParisC.Bootstrap_storage

This module provides functions that can be used in a private network to delay initial rewarding, typically when waiting for more bakers to join the network.

val init : Raw_context.t -> typecheck_smart_contract: (Raw_context.t -> Script_repr.t -> ((Script_repr.t * Lazy_storage_diff.diffs option) * Raw_context.t) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t) -> typecheck_smart_rollup: (Raw_context.t -> Script_repr.expr -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult) -> ?no_reward_cycles:int -> Parameters_repr.bootstrap_account list -> Parameters_repr.bootstrap_contract list -> Parameters_repr.bootstrap_smart_rollup list -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val cycle_end : Raw_context.t -> Cycle_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t