Module Tezos_raw_protocol_020_PsParisC.Commitment_storage

val exists : Raw_context.t -> Blinded_public_key_hash.t -> bool Tezos_protocol_environment_020_PsParisC.Lwt.t

exists ctxt bpkh returns true iff bpkh is associated to a non null commitment.

val committed_amount : Raw_context.t -> Blinded_public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

committed_amount ctxt bpkh return the commitment associated to bpkh, or Tez_repr.zero if bpkh has no associated commitment.

val increase_commitment_only_call_from_token : Raw_context.t -> Blinded_public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val decrease_commitment_only_call_from_token : Raw_context.t -> Blinded_public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val fold : Raw_context.t -> order:[ `Sorted | `Undefined ] -> init:'a -> f: (Blinded_public_key_hash.t -> Tez_repr.t -> 'a -> 'a Tezos_protocol_environment_020_PsParisC.Lwt.t) -> 'a Tezos_protocol_environment_020_PsParisC.Lwt.t