Module Contract_repr.Index

type t = t

include Path_encoding.S with type t := t

val to_path : t -> string list -> string list

to_path t postfix returns the context path name for t postfixed with postfix

val of_path : string list -> t option

of_path path parses path as a context path name for t

val path_length : int

Directory levels of the path encoding of t

val rpc_arg : t Tezos_protocol_environment_020_PsParisC.RPC_arg.t

val encoding : t Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val compare : t -> t -> int