Module Dal_operations_repr.Publish_commitment

type t = {

slot_index : Dal_slot_index_repr.t;

commitment : Dal_slot_repr.Commitment.t;

commitment_proof : Dal_slot_repr.Commitment_proof.t;

}

A "publish slot header" operation contains

a slot_index which is the slot index associated with the commitment.

a commitment which is a commitment to the slot data published onto the DAL

a commitment_proof which aims to prove that the size of the slot data does not exceed a limit set by the protocol.

val encoding : t Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val pp : Tezos_protocol_environment_020_PsParisC.Format.formatter -> t -> unit

val slot_header : cryptobox:Tezos_protocol_environment_020_PsParisC.Dal.t -> number_of_slots:int -> current_level:Raw_level_repr.t -> t -> Dal_slot_repr.Header.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

slot_header ~cryptobox ~number_of_slots ~current_level operation constructs a valid slot header. This function can fail in the following cases:

The published_level is not equal to current_level

The commitment_proof is invalid

The slot_index is invalid