Module Delegate_storage.For_RPC

The functions in this module are considered too costly to be used in the protocol. They are meant to be used only to answer RPC calls.

val full_balance : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

Returns the full 'balance' of the implicit contract associated to a given key, i.e. the sum of the spendable balance (given by balance or Contract_storage.get_balance) and of the frozen balance. The frozen balance is composed of all frozen bonds associated to the contract (given by Contract_storage.get_frozen_bonds) and of the part of the frozen deposits (given by frozen_deposits) that belongs to the delegate.

Only use this function for RPCs: this is expensive.

val delegated_balance : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

Only use this function for RPCs: this is expensive.

val staking_balance : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val min_delegated_in_current_cycle : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> (Tez_repr.t * Level_repr.t option) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t