Module Raw_context.Dal

type cryptobox = Tezos_protocol_environment_020_PsParisC.Dal.t

val make : t -> (t * cryptobox) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

val number_of_slots : t -> int

val number_of_shards : t -> int

val record_number_of_attested_shards : t -> Dal_attestation_repr.t -> int -> t

record_number_of_attested_shards ctxt attestation number_of_shards records that the number_of_shards shards were attested (declared available by some attester).

val register_slot_header : t -> Dal_slot_repr.Header.t -> t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

register_slot_header ctxt slot_header returns a new context where the new candidate slot have been taken into account. Returns Some (ctxt,updated) where updated=true if the candidate is registered. Some (ctxt,false) if another candidate was already registered previously. Returns an error if the slot is invalid.

val candidates : t -> Dal_slot_repr.Header.t list

candidates ctxt returns the current list of slot for which there is at least one candidate.

val is_slot_index_attested : t -> Dal_slot_index_repr.t -> bool

is_slot_index_attested ctxt slot_index returns true if the slot_index is declared available by the protocol. false otherwise. If the index is out of the interval 0;number_of_slots - 1, returns false.