Module Tezos_raw_protocol_020_PsParisC.Sampler

Efficient sampling from given finitely supported (nonzero, positive) measures using the alias method. Measures need not be normalized on input, but sampling proceeds from the normalized probability measure associated to the given measure.

module type S = sig ... end

S is the module type of a module allowing to construct samplers based on the alias method.