Module Sc_rollup_management_protocol.Internal_for_tests

Function for constructing and encoding inbox_message and outbox_message values. Since Layer 1 only ever consumes outbox_message values and produces inbox_message values, these functions are used for testing only.

val make_transaction : Alpha_context.context -> ('a, _) Script_typed_ir.ty -> parameters:'a -> destination:Contract_hash.t -> entrypoint:Alpha_context.Entrypoint.t -> (transaction * Alpha_context.context) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

make_transaction ctxt ty ~parameters ~destination ~entrypoint creates a Layer 1 to Layer 2 transaction.

val make_atomic_batch : transaction list -> outbox_message

make_atomic_batch ts creates an atomic batch with the given transactions ts.

val serialize_outbox_transactions_untyped : transaction list -> Alpha_context.Sc_rollup.Outbox.Message.serialized Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

serialize_output_transactions_untyped t encodes the outbox transaction batch t in binary format using the untyped outbox message representation.

val serialize_outbox_transactions_typed : transaction list -> Alpha_context.Sc_rollup.Outbox.Message.serialized Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

serialize_output_transactions_typed t encodes the outbox transaction batch t in binary format using the typed outbox message representation.

val deserialize_inbox_message : Alpha_context.Sc_rollup.Inbox_message.serialized -> Alpha_context.Sc_rollup.Inbox_message.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

deserialize_inbox_message bs decodes an inbox message from the given data bs.