Module Tezos_raw_protocol_020_PsParisC.Sc_rollup_operations

High-level operations over smart contract rollups.

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Sc_rollup_invalid_parameters_type

| Sc_rollup_invalid_last_cemented_commitment

| Sc_rollup_invalid_output_proof

| Sc_rollup_invalid_outbox_level

type execute_outbox_message_result = {

paid_storage_size_diff : Tezos_protocol_environment_020_PsParisC.Z.t;

ticket_receipt : Ticket_receipt.t;

operations : Script_typed_ir.packed_internal_operation list;

whitelist_update : Alpha_context.Sc_rollup.Whitelist.update option;

}

Result of calling the execute_outbox_message function.

type origination_result = {

address : Alpha_context.Sc_rollup.Address.t;

size : Tezos_protocol_environment_020_PsParisC.Z.t;

genesis_commitment_hash : Alpha_context.Sc_rollup.Commitment.Hash.t;

}

val originate : ?whitelist:Alpha_context.Sc_rollup.Whitelist.t -> Alpha_context.context -> kind:Alpha_context.Sc_rollup.Kind.t -> boot_sector:string -> parameters_ty:Script_repr.lazy_expr -> (origination_result * Alpha_context.context) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

originate ?whitelist context ~kind ~boot_sector ~parameters_ty adds a new rollup running in a given kind initialized with a boot_sector and to accept smart contract calls of type parameters_ty.

val execute_outbox_message : Alpha_context.context -> Alpha_context.Sc_rollup.t -> cemented_commitment:Alpha_context.Sc_rollup.Commitment.Hash.t -> output_proof:string -> (execute_outbox_message_result * Alpha_context.context) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

execute_outbox_message ctxt rollup ~cemented_commitment ~output_proof validates the given outbox message and prepares a set of resulting operations.

val validate_untyped_parameters_ty : Alpha_context.context -> Alpha_context.Script.expr -> Alpha_context.context Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

validate_untyped_parameters_ty ctxt script parses the type and check that the entrypoints are well-formed.

module Internal_for_tests : sig ... end

A module used for testing purposes only.