Module Sc_rollup_wasm.V2_0_0

This module provides Proof-Generating Virtual Machine (PVM) running WebAssembly (version 2.0.0).

val current_version : Tezos_protocol_environment_020_PsParisC.Wasm_2_0_0.version

module type S = sig ... end

module type Make_wasm = module type of Tezos_protocol_environment_020_PsParisC.Wasm_2_0_0.Make

module Make (Lib_scoru_Wasm : Make_wasm) (Context : Sc_rollup_PVM_sig.Generic_pvm_context_sig) : S with type context = Context.Tree.t and type state = Context.tree and type proof = Context.proof

Build a WebAssembly PVM using the given proof-supporting context.

module Protocol_implementation : S with type context = Tezos_protocol_environment_020_PsParisC.Context.t and type state = Tezos_protocol_environment_020_PsParisC.Context.tree and type proof = Tezos_protocol_environment_020_PsParisC.Context.Proof.tree Tezos_protocol_environment_020_PsParisC.Context.Proof.t

This PVM is used for verification in the Protocol. produce_proof always returns None.

val ticks_per_snapshot : Tezos_protocol_environment_020_PsParisC.Z.t

Number of ticks between snapshotable states, chosen low enough to maintain refutability.

Warning: This value is used to specialize the dissection predicate of the WASM PVM. Do not change it without a migration plan for already originated smart rollups.

Depends on

speed (tick/s) of node in slow mode (from benchmark, 6000000 tick/s)

the number of ticks in a commitment (Int64.max_int, as per Number_of_ticks.max_value)

see #3590 for more pointers

val outbox_validity_period : int32

val outbox_message_limit : Tezos_protocol_environment_020_PsParisC.Z.t

val well_known_reveal_hash : Sc_rollup_reveal_hash.t

The hash requested by the WASM PVM if it cannot decode the input provided by the WASM kernel, that is, if the bytes value cannot be decoded with Sc_rollup_reveal_hash.encoding.

val well_known_reveal_preimage : string

The preimage of well_known_reveal_hash.

val decode_reveal : Tezos_protocol_environment_020_PsParisC.Wasm_2_0_0.reveal -> Sc_rollup_PVM_sig.reveal

Convert a raw reveal request of the WASM PVM into a typed reveal as defined by the protocol.

If the decoding fails, fallback to requesting the preimage of the well_known_reveal_hash.