Module Tezos_raw_protocol_020_PsParisC.Script_tc_errors

type kind =

| Int_kind

| String_kind

| Bytes_kind

| Prim_kind

| Seq_kind

type unparsed_stack_ty = Alpha_context.Script.expr list

type type_map = (Alpha_context.Script.location * (unparsed_stack_ty * unparsed_stack_ty)) list

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_arity of Alpha_context.Script.location * Alpha_context.Script.prim * int * int

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_seq_arity of Alpha_context.Script.location * int * int

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_namespace of Alpha_context.Script.location * Alpha_context.Script.prim * Michelson_v1_primitives.namespace * Michelson_v1_primitives.namespace

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_primitive of Alpha_context.Script.location * Alpha_context.Script.prim list * Alpha_context.Script.prim

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_kind of Alpha_context.Script.location * kind list * kind

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_never_expr of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Missing_field of Alpha_context.Script.prim

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Duplicate_field of Alpha_context.Script.location * Alpha_context.Script.prim

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_lazy_storage of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_operation of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_contract of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| No_such_entrypoint of Alpha_context.Entrypoint.t

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Duplicate_entrypoint of Alpha_context.Entrypoint.t

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unreachable_entrypoint of Alpha_context.Script.prim list

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Tx_rollup_bad_deposit_parameter of Alpha_context.Script.location * Alpha_context.Script.expr

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Tx_rollup_invalid_ticket_amount of Tezos_protocol_environment_020_PsParisC.Z.t

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Forbidden_zero_ticket_quantity

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Sc_rollup_disabled of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Zk_rollup_disabled of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Zk_rollup_bad_deposit_parameter of Alpha_context.Script.location * Alpha_context.Script.expr

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Fail_not_in_tail_position of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Undefined_binop : Alpha_context.Script.location * Alpha_context.Script.prim * Alpha_context.Script.expr * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Undefined_unop : Alpha_context.Script.location * Alpha_context.Script.prim * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Bad_return : Alpha_context.Script.location * unparsed_stack_ty * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Bad_stack : Alpha_context.Script.location * Alpha_context.Script.prim * int * unparsed_stack_ty -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unmatched_branches : Alpha_context.Script.location * unparsed_stack_ty * unparsed_stack_ty -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| View_name_too_long of string

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Bad_view_name of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Ill_typed_view of {

loc : Alpha_context.Script.location;

actual : unparsed_stack_ty;

expected : unparsed_stack_ty;

}

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Duplicated_view_name of Alpha_context.Script.location

type context_desc =

| Lambda

| View

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Forbidden_instr_in_context of Alpha_context.Script.location * context_desc * Alpha_context.Script.prim

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Bad_stack_length

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Bad_stack_item of int

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_annotation of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Ungrouped_annotations of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_map_body : Alpha_context.Script.location * unparsed_stack_ty -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_map_block_fail of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_iter_body : Alpha_context.Script.location * unparsed_stack_ty * unparsed_stack_ty -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Type_too_large : Alpha_context.Script.location * int -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Pair_bad_argument of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unpair_bad_argument of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Dup_n_bad_argument of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Dup_n_bad_stack of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_constant : Alpha_context.Script.location * Alpha_context.Script.expr * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_syntactic_constant : Alpha_context.Script.location * Alpha_context.Script.expr * string -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_contract of Alpha_context.Script.location * Alpha_context.Contract.t

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Invalid_big_map of Alpha_context.Script.location * Alpha_context.Big_map.Id.t

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Comparable_type_expected : Alpha_context.Script.location * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Inconsistent_type_sizes : int * int -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Inconsistent_types : Alpha_context.Script.location * Alpha_context.Script.expr * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_implicit_account_parameters_type : Alpha_context.Script.location * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Inconsistent_memo_sizes : Alpha_context.Sapling.Memo_size.t * Alpha_context.Sapling.Memo_size.t -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unordered_map_keys of Alpha_context.Script.location * Alpha_context.Script.expr

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unordered_set_values of Alpha_context.Script.location * Alpha_context.Script.expr

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Duplicate_map_keys of Alpha_context.Script.location * Alpha_context.Script.expr

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Duplicate_set_values of Alpha_context.Script.location * Alpha_context.Script.expr

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Ill_typed_data : string option * Alpha_context.Script.expr * Alpha_context.Script.expr -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Ill_formed_type of string option * Alpha_context.Script.expr * Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Ill_typed_contract : Alpha_context.Script.expr * type_map -> Tezos_protocol_environment_020_PsParisC.Error_monad.error

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Deprecated_instruction of Alpha_context.Script.prim

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Typechecking_too_many_recursive_calls

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unparsing_too_many_recursive_calls

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_ticket of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_forged_value of Alpha_context.Script.location

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Non_dupable_type of Alpha_context.Script.location * Alpha_context.Script.expr

type Tezos_protocol_environment_020_PsParisC.Error_monad.error +=

| Unexpected_ticket_owner of Alpha_context.Destination.t

type inconsistent_types_fast_error =

| Inconsistent_types_fast
(*
This value is only used when the details of the error don't matter because the error will be ignored later. For example, when types are compared during the interpretation of the CONTRACT instruction any error will lead to returning None but the content of the error will be ignored.
*)

type (_, _) error_details =

| Informative : 'error_context -> ('error_context, Tezos_protocol_environment_020_PsParisC.Error_monad.error Tezos_protocol_environment_020_PsParisC.Error_monad.trace) error_details

| Fast : (_, inconsistent_types_fast_error) error_details