Module Tezos_raw_protocol_020_PsParisC.Seed_repr

Tezos Protocol Implementation - Random number generation

This is not expected to be a good cryptographic random number generator. In particular this is supposed to be used in situations where the seed is a globally known information.

The only expected property is: It should be difficult to find a seed such that the generated sequence is a given one.

Random Generation

type seed

A random seed, to derive random sequences from

type vdf_setup = Tezos_protocol_environment_020_PsParisC.Vdf.discriminant * Tezos_protocol_environment_020_PsParisC.Vdf.challenge

A VDF discriminant and challenge

type vdf_solution = Tezos_protocol_environment_020_PsParisC.Vdf.result * Tezos_protocol_environment_020_PsParisC.Vdf.proof

A VDF result, to derive a seed from

val pp_solution : Tezos_protocol_environment_020_PsParisC.Format.formatter -> vdf_solution -> unit

val compare_vdf_solution : vdf_solution -> vdf_solution -> int

Compare only the first element of two vdf_solution, that are of Vdf.result.

val generate_vdf_setup : seed_discriminant:seed -> seed_challenge:seed -> vdf_setup

val verify : vdf_setup -> Tezos_protocol_environment_020_PsParisC.Int64.t -> vdf_solution -> bool option

val vdf_to_seed : seed -> vdf_solution -> seed

Entropy

type nonce

A nonce for adding entropy to the generator

val update_seed : seed -> nonce -> seed

Add entropy to the seed generator

val make_nonce : bytes -> nonce Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

Use a byte sequence as a nonce

val hash : nonce -> Nonce_hash.t

Compute the hash of a nonce

val check_hash : nonce -> Nonce_hash.t -> bool

check_hash nonce hash is true if the nonce correspond to the hash

val nonce_hash_key_part : Nonce_hash.t -> string list -> string list

For using nonce hashes as keys in the hierarchical database

val deterministic_seed : seed -> seed

Returns a new seed by hashing the one passed with a constant.

val initial_seeds : ?initial_seed:State_hash.t -> int -> seed list

initial_seeds n generates the first n seeds for which there are no nonces. The first seed is a constant value. The kth seed is the hash of seed (k-1) concatenated with a constant. If an initial_seed is provided, the first seed is created using it as the first one.

Serializers

val nonce_encoding : nonce Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val seed_encoding : seed Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val vdf_setup_encoding : vdf_setup Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val vdf_solution_encoding : vdf_solution Tezos_protocol_environment_020_PsParisC.Data_encoding.t

type seed_status =

| RANDAO_seed

| VDF_seed

val seed_status_encoding : seed_status Tezos_protocol_environment_020_PsParisC.Data_encoding.t