Module Tezos_raw_protocol_020_PsParisC.Shared_stake

type shared = {

baker_part : Tez_repr.t;

stakers_part : Tez_repr.t;

}

val share : rounding:[ `Towards_stakers | `Towards_baker ] -> Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.public_key_hash -> Tez_repr.t -> shared Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

share ~rounding ctxt delegate amount shares amount between a baker_part and a stakers_part proportionally to their shares in delegate's frozen deposits. rounding controls where the extra mutez goes in case the split doesn't fall on integer values.

val pay_rewards : Raw_context.t -> ?active_stake:Stake_repr.t -> source:[< Token.giver ] -> delegate:Tezos_protocol_environment_020_PsParisC.Signature.public_key_hash -> Tez_repr.t -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

pay_rewards ctxt ?active_stake source delegate transfers the rewards to the delegate spendable balance and frozen balance.

The distribution is based on the baker's staking parameters.

If adaptive issuance is enabled, it also accounts for the repartition of the delegate's active_stake between delegated token and frozen deposits. If active_stake is not provided, it will be retrieved from the context.