Module Tezos_raw_protocol_020_PsParisC.Slash_percentage

val get : Raw_context.t -> kind:Misbehaviour_repr.kind -> level:Level_repr.t -> Tezos_protocol_environment_020_PsParisC.Signature.public_key_hash list -> (Raw_context.t * Percentage.t) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

get ctxt ~kind ~level denounced returns the percentage that needs to be applied for the given misbehaviour.

denounced is the list of delegates that have been denounced together for the given kind, for the given level and for the same round. The amount slashed increases quadratically as the number of attesting slots of denounced delegates increases. The maximum slashing value max_slashing_per_block is reached when that number of slots reaches max_slashing_threshold .

module Internal_for_tests : sig ... end