Module Tezos_raw_protocol_020_PsParisC.Stake_storage

This module provides basic operations (accessors and setters) on staking tokens.

It is responsible for maintaining the following tables:

Storage.Stake.Selected_distribution_for_cycle

Storage.Stake.Staking_balance

Storage.Stake.Active_delegates_with_minimal_stake

Storage.Stake.Total_active_stake

val initialize_delegate : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.public_key_hash -> delegated:Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val get_full_staking_balance : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.public_key_hash -> Full_staking_balance_repr.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val remove_delegated_stake : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val remove_frozen_stake_only_call_from_token : Raw_context.t -> Frozen_staker_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val add_delegated_stake : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val add_frozen_stake_only_call_from_token : Raw_context.t -> Frozen_staker_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val set_inactive : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Lwt.t

val set_active : Raw_context.t -> Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val fold_on_active_delegates_with_minimal_stake_es : Raw_context.t -> f: (Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> 'a -> 'a Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t) -> order:[ `Sorted | `Undefined ] -> init:'a -> 'a Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

fold ctxt ~f ~order init folds f on the list of active delegates having the minimal required stake. The folding process starts with init. Each element of the list is the public key hash of a delegate.

val set_selected_distribution_for_cycle : Raw_context.t -> Cycle_repr.t -> (Tezos_protocol_environment_020_PsParisC.Signature.public_key_hash * Stake_repr.t) list -> Stake_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

set_selected_distribution_for_cycle ctxt cycle distrib total_stake saves the selected distribution distrib of the total_stake for the given cycle.

val clear_at_cycle_end : Raw_context.t -> new_cycle:Cycle_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val fold_on_active_delegates_with_minimal_stake_s : Raw_context.t -> order:[ `Sorted | `Undefined ] -> init:'a -> f: (Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t -> 'a -> 'a Tezos_protocol_environment_020_PsParisC.Lwt.t) -> 'a Tezos_protocol_environment_020_PsParisC.Lwt.t

val get_selected_distribution : Raw_context.t -> Cycle_repr.t -> (Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t * Stake_repr.t) list Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val find_selected_distribution : Raw_context.t -> Cycle_repr.t -> (Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.t * Stake_repr.t) list option Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val get_selected_distribution_as_map : Raw_context.t -> Cycle_repr.t -> Stake_repr.t Tezos_protocol_environment_020_PsParisC.Signature.Public_key_hash.Map.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val prepare_stake_distribution : Raw_context.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

Copy the stake distribution for the current cycle (from Storage.Stake.Selected_distribution_for_cycle) in the raw context.

val get_total_active_stake : Raw_context.t -> Cycle_repr.t -> Stake_repr.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

get_total_active_stake ctxt cycle retrieves the amount in Tez of the active stake at cycle from ctxt.

val add_contract_delegated_stake : Raw_context.t -> Contract_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

add_contract_delegated_stake ctxt contract amount calls Stake_storage.add_delegated_stake ctxt delegate amount if contract has a delegate. Otherwise this function does nothing.

val remove_contract_delegated_stake : Raw_context.t -> Contract_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

remove_contract_delegated_stake ctxt contract amount calls Stake_storage.remove_delegated_stake ctxt delegate amount if contract has a delegate. Otherwise this function does nothing.

val cleanup_values_for_protocol_p : Raw_context.t -> preserved_cycles:int -> consensus_rights_delay:int -> new_cycle:Cycle_repr.t -> Raw_context.t Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

module For_RPC : sig ... end

module Internal_for_tests : sig ... end