Module Storage.Adaptive_issuance

module Launch_ema : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Tezos_protocol_environment_020_PsParisC.Int32.t

Exponential moving average (ema) of votes set in the block header protocol_data.contents. Once the feature is activated, it can no longer be deactivated without a protocol amendment. *

module Activation : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Cycle_repr.t option

Cycle Some c from which adaptive issuance is (or will be) active, or None if the feature is not yet planned to activate. *