Module Dal.Slot

module Headers : Storage_sigs.Non_iterable_indexed_data_storage with type t = Raw_context.t and type key = Raw_level_repr.t and type value = Dal_slot_repr.Header.t list

This is a temporary storage for slot headers proposed onto the L1.

module History : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Dal_slot_repr.History.t

This is a permanent storage for slot headers confirmed by the L1.