Module Tezos_raw_protocol_020_PsParisC.Ticket_balance_key

This module exposes a function for generating a ticket-balance key-hash given an owner and a ticket-token. The key-hash is used for populating the global ticket-balance table that tracks ownership of tickets for different tokens.

val make : Alpha_context.context -> owner:Alpha_context.Destination.t -> ticketer:Alpha_context.Contract.t -> contents_type:Alpha_context.Script.node -> contents:Alpha_context.Script.node -> (Alpha_context.Ticket_hash.t * Alpha_context.context) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

make ~owner ~ticketer ~contents_type ~contents returns key_hash of the given owner, ticketer, contents_type and contents. Note that the location of the Script.node values contents_type and contents are irrelevant since Ticket_hash.make will strip the locations before calculating the hash.

val of_ex_token : Alpha_context.context -> owner:Alpha_context.Destination.t -> Ticket_token.ex_token -> (Alpha_context.Ticket_hash.t * Alpha_context.context) Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

of_ex_token ctxt ~owner ex_token returns the key_hash of the given owner and ex_token.