Module Tezos_raw_protocol_020_PsParisC.Ticket_hash_repr

Ticket hashes are used to uniquely identify pairs made of Michelson ticktes and their owner.

They are used by the protocol to keep record of a tickets ledger, that is how many tickets smart contracts own. More precisely, they are used as keys for the Storage.Ticket_balance table.

type t

A ticket hash is computed by the function make and is a combination of a ticketer, a content type, a content, and an owner.

Note: This invariant can be invalidated if the key_hash is created from the encoding.

val encoding : t Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val pp : Tezos_protocol_environment_020_PsParisC.Format.formatter -> t -> unit

val to_b58check : t -> string

val of_b58check_opt : string -> t option

val of_b58check_exn : string -> t

val of_bytes_exn : bytes -> t

val of_bytes_opt : bytes -> t option

include Tezos_protocol_environment_020_PsParisC.Compare.S with type t := t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val zero : t

val of_script_expr_hash : Script_expr_hash.t -> t

module Index : Storage_description.INDEX with type t = t